+++ lib/carStaticParams.ts
@@ -1,5 +1,9 @@
 import cars from "@/data/cars.json";
 
+type CatalogCar = {
+  lot_id: string;
+};
+
 export function carRouteStaticParams() {
-  return cars.map((car) => ({ id: car.lot_id }));
+  return (cars as CatalogCar[]).map((car) => ({ id: car.lot_id }));
 }